backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-generativeai>=0.8.3
python-dotenv>=1.0.1
redis>=5.2.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
